requirements.txt
trafilatura
aiohttp
telegraph
beautifulsoup4
python-dotenv
google-genai
python-telegram-bot
packaging
markdown-it-py
telegramify-markdown
